feat: :art: The HTML structure is built and the JavaScript file is created
The HTML file structure is implemented, including a background image for the page, a heading displaying the veterinary clinic's name, and two buttons. Each button is intended to trigger separate functionalities that comply with the requirements specified in the assignment. These buttons will later be linked to specific JavaScript functions to be developed

by@Alexis-Blanco