feat: :sparkles: The function to register a new owner is created.
The function to register an owner is created. It prompts for the owner's name, ID number, phone number, and email. A variable is also implemented to verify that the ID number is unique; if the ID is not duplicated, the new owner is created and added to the previously created owners list.

by@Alexis-Blanco